LayoutTests/fast/mediastream/audio-settings-reset.html

<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
promise_test(async test => {
    const stream1 = await navigator.mediaDevices.getUserMedia({audio : true });
    test.add_cleanup(() => stream1.getTracks().forEach(track => track.stop()));
    assert_equals(stream1.getAudioTracks()[0].getSettings().volume, 1, "test1");

    const stream2 = await navigator.mediaDevices.getUserMedia({audio : { volume : 0 } });
    test.add_cleanup(() => stream2.getTracks().forEach(track => track.stop()));
    assert_equals(stream2.getAudioTracks()[0].getSettings().volume, 0);

    const stream3 = await navigator.mediaDevices.getUserMedia({audio : true });
    test.add_cleanup(() => stream3.getTracks().forEach(track => track.stop()));
    assert_equals(stream3.getAudioTracks()[0].getSettings().volume, 1, "test3");
}, "Volume should be reset to default value if not specified explicitly");
</script>
</body>
